refactor: move remaining lib modules to util, since their roles are identical to util
