+++ app/create/records/page.tsx
@@ -4,7 +4,7 @@ import { redirect } from 'next/navigation';
 import { getArcadeInfoList } from '^/src/features/arcade-info/data';
 import RecordForm from '^/src/features/arcade-record-article/record-form';
 import { getMethodList } from '^/src/features/method/data';
-import { APP_NAME } from '^/src/shared/lib/is-production';
+import { APP_NAME } from '^/src/shared/util/is-production';
 import { createServerSideClient } from '^/src/shared/supabase/server';
 
 export const metadata: Metadata = {
