+++ app/gallery/page.tsx
@@ -5,7 +5,7 @@ import EmptySvg from '^/public/status/empty.svg';
 import { GALLERY_PHOTOS_PER_PAGE } from '^/src/entities/constants/pagenation';
 import GalleryPostList from '^/src/features/gallery';
 import { getGalleryList } from '^/src/features/gallery/data';
-import { APP_NAME } from '^/src/shared/lib/is-production';
+import { APP_NAME } from '^/src/shared/util/is-production';
 
 export const metadata: Metadata = {
   title: `갤러리 :: ${APP_NAME}`,
